Been debugging the project to get rid of a bug present
there is a bug that is lingering in the code that I need to get rid of.

Im going to install editor symbols to get to the bottom of the bug and find the root cause of the issue.
Its likely going to be something simple to do with Garbage collection, but I cannot find it yet.
Hopefully the editor symbols plugin can help